refactor: trying to fix bug with prisma module
