+++ package.json
@@ -57,7 +57,7 @@
     "@types/express": "^4.17.13",
     "@types/jest": "27.4.1",
     "@types/jsonwebtoken": "^8.5.8",
-    "@types/node": "^16.0.0",
+    "@types/node": "^16.18.12",
     "@types/passport-jwt": "^3.0.8",
     "@types/passport-local": "^1.0.35",
     "@types/supertest": "^2.0.11",
@@ -75,9 +75,9 @@
     "supertest": "^6.2.3",
     "ts-jest": "^27.0.3",
     "ts-loader": "^9.2.3",
-    "ts-node": "^10.0.0",
+    "ts-node": "^10.9.1",
     "tsconfig-paths": "^3.10.1",
-    "typescript": "^4.3.5"
+    "typescript": "^4.9.5"
   },
   "jest": {
     "moduleFileExtensions": [
