+++ src/album/album.module.ts
@@ -1,11 +1,10 @@
 import { Module } from '@nestjs/common';
-import { PrismaModule } from 'src/prisma.module';
+import { PrismaService } from 'src/prisma/prisma.service';
 import { AlbumController } from './album.controller';
 import { AlbumService } from './album.service';
 
 @Module({
   controllers: [AlbumController],
-  providers: [AlbumService],
-  imports: [PrismaModule],
+  providers: [AlbumService, PrismaService],
 })
 export class AlbumModule {}
