+++ src/app.module.ts
@@ -3,20 +3,23 @@ import { AlbumModule } from './album/album.module';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { ArtistModule } from './artist/artist.module';
+import { AuthModule } from './auth/auth.module';
 import { FavsModule } from './favs/favs.module';
-import { PrismaModule } from './prisma.module';
+import { PrismaModule } from './prisma/prisma.module';
 import { TrackModule } from './track/track.module';
 import { UserModule } from './user/user.module';
-import { AuthModule } from './auth/auth.module';
 
 @Module({
   imports: [
+    // ConfigModule.forRoot({
+    //   isGlobal: true,
+    // }),
+    PrismaModule,
     UserModule,
     ArtistModule,
     TrackModule,
     AlbumModule,
     FavsModule,
-    PrismaModule,
     AuthModule,
   ],
   controllers: [AppController],
