+++ src/artist/artist.module.ts
@@ -1,11 +1,11 @@
 import { Module } from '@nestjs/common';
-import { PrismaModule } from 'src/prisma.module';
+import { PrismaService } from 'src/prisma/prisma.service';
 import { ArtistController } from './artist.controller';
 import { ArtistService } from './artist.service';
 
 @Module({
   controllers: [ArtistController],
-  providers: [ArtistService],
-  imports: [PrismaModule],
+  providers: [ArtistService, PrismaService],
+  // imports: [PrismaModule],
 })
 export class ArtistModule {}
