+++ src/auth/auth.service.ts
@@ -1,11 +1,10 @@
 import { BadRequestException, Injectable } from '@nestjs/common';
-import { ConfigService } from '@nestjs/config';
 import { JwtService } from '@nestjs/jwt';
 import * as bcrypt from 'bcrypt';
 import { CRYPT_SALT } from 'src/main';
-import { PrismaService } from 'src/prisma.service';
+import { PrismaService } from 'src/prisma/prisma.service';
 import { UserService } from 'src/user/user.service';
-import { ERROR_MESSAGES } from 'utils/constants';
+import { ERROR_MESSAGES } from 'src/utils/constants';
 import { LoginAuthDto } from './dto/login-auth.dto';
 import { RefreshAuthDto } from './dto/refresh-auth.dto';
 import { SignupAuthDto } from './dto/signup-auth.dto';
@@ -15,7 +14,7 @@ export class AuthService {
   constructor(
     private userService: UserService,
     private jwtService: JwtService,
-    private configService: ConfigService,
+    // private configService: ConfigService,
     private prisma: PrismaService,
   ) {}
 
@@ -29,13 +28,10 @@ export class AuthService {
     }
 
     const hash = await this.hashData(password);
-    const newUser = await this.userService.create({
+    return await this.userService.create({
       ...signupAuthDto,
       password: hash,
     });
-    const tokens = await this.getTokens(newUser._id, newUser.login);
-    await this.updateRefreshToken(newUser._id, tokens.refreshToken);
-    return tokens;
   }
 
   async login(loginAuthDto: LoginAuthDto) {
@@ -50,39 +46,39 @@ export class AuthService {
 
   async updateRefreshToken(id: string, refreshToken: string) {
     const hashedRefreshToken = await this.hashData(refreshToken);
-    await this.prisma.user.update({
-      where: { id },
-      data: { refreshToken: hashedRefreshToken },
-    });
+    // await this.prisma.user.update({
+    //   where: { id },
+    //   data: { refreshToken: hashedRefreshToken },
+    // });
   }
 
-  async getTokens(userId: string, login: string) {
-    const [accessToken, refreshToken] = await Promise.all([
-      this.jwtService.signAsync(
-        {
-          sub: userId,
-          login,
-        },
-        {
-          secret: this.configService.get<string>('JWT_ACCESS_SECRET'),
-          expiresIn: '15m',
-        },
-      ),
-      this.jwtService.signAsync(
-        {
-          sub: userId,
-          login,
-        },
-        {
-          secret: this.configService.get<string>('JWT_REFRESH_SECRET'),
-          expiresIn: '7d',
-        },
-      ),
-    ]);
+  // async getTokens(userId: string, login: string) {
+  //   const [accessToken, refreshToken] = await Promise.all([
+  //     this.jwtService.signAsync(
+  //       {
+  //         sub: userId,
+  //         login,
+  //       },
+  //       {
+  //         secret: this.configService.get<string>('JWT_ACCESS_SECRET'),
+  //         expiresIn: '15m',
+  //       },
+  //     ),
+  //     this.jwtService.signAsync(
+  //       {
+  //         sub: userId,
+  //         login,
+  //       },
+  //       {
+  //         secret: this.configService.get<string>('JWT_REFRESH_SECRET'),
+  //         expiresIn: '7d',
+  //       },
+  //     ),
+  //   ]);
 
-    return {
-      accessToken,
-      refreshToken,
-    };
-  }
+  //   return {
+  //     accessToken,
+  //     refreshToken,
+  //   };
+  // }
 }
