+++ src/favs/favs.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { PrismaModule } from 'src/prisma.module';
+import { PrismaModule } from 'src/prisma/prisma.module';
 import { FavsController } from './favs.controller';
 import { FavsService } from './favs.service';
