+++ src/favs/favs.service.ts
@@ -3,8 +3,8 @@ import {
   NotFoundException,
   UnprocessableEntityException,
 } from '@nestjs/common';
-import { PrismaService } from 'src/prisma.service';
-import { ERROR_MESSAGES } from 'utils/constants';
+import { PrismaService } from 'src/prisma/prisma.service';
+import { ERROR_MESSAGES } from 'src/utils/constants';
 
 @Injectable()
 export class FavsService {
