+++ src/track/track.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { PrismaModule } from 'src/prisma.module';
+import { PrismaModule } from 'src/prisma/prisma.module';
 import { TrackController } from './track.controller';
 import { TrackService } from './track.service';
